chore: changes in formatting and cleanup
