+++ efemel/main.py
@@ -4,14 +4,14 @@ Main module for the efemel project.
 
 
 def hello_world():
-    """
-    Simple hello world function.
+  """
+  Simple hello world function.
 
-    Returns:
-        str: A greeting message
-    """
-    return "Hello, World from efemel!"
+  Returns:
+      str: A greeting message
+  """
+  return "Hello, World from efemel!"
 
 
 if __name__ == "__main__":
-    print(hello_world())
+  print(hello_world())
