--- output_recursive/test_dir/config.json
@@ -1,10 +1,0 @@
-{
-  "app_config": {
-    "app_name": "my_app",
-    "port": 8080
-  },
-  "database_config": {
-    "url": "postgresql://localhost/mydb",
-    "pool_size": 10
-  }
-}
