--- output_recursive/test_dir/subdir/nested.json
@@ -1,14 +1,0 @@
-{
-  "metadata": {
-    "created_by": "test",
-    "timestamp": "2025-01-01"
-  },
-  "nested_data": {
-    "level": "deep",
-    "items": [
-      1,
-      2,
-      3
-    ]
-  }
-}
