+++ pyproject.toml
@@ -31,6 +31,7 @@ dev = [
 [tool.ruff]
 line-length = 88
 target-version = "py312"
+indent-width = 2
 
 [tool.ruff.lint]
 select = [
